run make generate

+++ apis/project/v1alpha1/zz_project_types.go
@@ -68,7 +68,7 @@ type ProjectObservation struct {
 
 	// the id of the project owner.
 	// the id of the project owner.
-	OwnerID *float64 `json:"ownerId,omitempty" tf:"owner_id,omitempty"`
+	OwnerID *int64 `json:"ownerId,omitempty" tf:"owner_id,omitempty"`
 
 	// the unique universal identifier of the project owner.
 	// the unique universal identifier of the project owner.
